requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
WTForms==3.0.1
Werkzeug==2.3.7
python-dotenv==1.0.0
bcrypt==4.0.1
stripe==5.5.0
razorpay==1.3.0
Pillow==10.4.0
email-validator==2.0.0